// src/app/interfaces/interfaces.ts
export interface PokemonDetails {
    dex?: number;
    name?: string;
    type?: string;
    height?: string;
    weight?: string;
    HP?: number;
    attack?: number;
    defense?: number;
    speed?: number;
    specialAttack?: number;
    specialDefense?: number;
    imageNormal?: string;
    imageShiny?: string;
}